MAT-9685: add health check endpoint outside of auth
