+++ src/export.module.ts
@@ -3,7 +3,7 @@ import { ExportController } from './controllers/export.controller';
 import { ExportService } from './services/export.service';
 import { AuthModule } from './auth/auth.module';
 import { TerminusModule } from '@nestjs/terminus';
-import { HealthController } from './health/health.controller';
+import { HealthController } from './controllers/health.controller';
 
 @Module({
   imports: [AuthModule, TerminusModule],
