--- src/health/health.controller.ts
@@ -1,13 +1,0 @@
-import { Controller, Get } from '@nestjs/common';
-import { HealthCheck, HealthCheckService } from '@nestjs/terminus';
-
-@Controller('health')
-export class HealthController {
-  constructor(private health: HealthCheckService) {}
-
-  @Get()
-  @HealthCheck()
-  check() {
-    return this.health.check([]);
-  }
-}
